+++ src/actions/fix.ts
@@ -1,0 +1,9 @@
+import chalk from 'chalk';
+import ora from 'ora';
+import { CachedESLint } from '../eslint';
+
+export async function doFixAction(eslint: CachedESLint, selectedRuleIds: string[]) {
+  const fixingSpinner = ora('Fixing...').start();
+  await eslint.fix(selectedRuleIds);
+  fixingSpinner.succeed(chalk.bold('Fixing was successful.'));
+}
